+++ altair/constants.js
@@ -14,8 +14,11 @@ const __NODE_ENV = process.env.NODE_ENV;
 const __APP_NAME = process.env.APP_NAME;
 const __ACTIVE_SPACE = process.env.ACTIVE_SPACE;
 const __PUBLIC_LOCATION = process.env.PUBLIC_LOCATION;
-const __PAGES_LOCATION = process.env.PAGES_LOCATION; 
+const __PAGES_LOCATION = process.env.PAGES_LOCATION;
+const __DATA_LOCATION = process.env.DATA_LOCATION || '';
+const __DATA_FILE = process.env.DATA_FILE || '';
+const __ENABLE_DATA_WATCH = process.env.ENABLE_DATA_WATCH !== 'false';
 const __MINIFY = process.env.MINIFY === 'true';
 const __DEBUG = process.env.DEBUG === 'true';
 
-export { __dirName, __appRootPath, __version, __localPort, __NODE_ENV, __APP_NAME, __ACTIVE_SPACE, __PUBLIC_LOCATION, __PAGES_LOCATION, __MINIFY, __DEBUG };
+export { __dirName, __appRootPath, __version, __localPort, __NODE_ENV, __APP_NAME, __ACTIVE_SPACE, __PUBLIC_LOCATION, __PAGES_LOCATION, __DATA_LOCATION, __DATA_FILE, __ENABLE_DATA_WATCH, __MINIFY, __DEBUG };